feat(analysis): add clickable metric explanations to multi-scenario analysis boxes
- Make all stat boxes (Success Rate, Median Balance, Percentiles, Expected Value, Std Deviation) clickable
- Add hover effects with border highlighting and subtle elevation
- Create detailed explanation modals for each metric covering:
  * What the metric means
  * How it's calculated
  * How to interpret the results
  * Real-world examples and analogies
  * Practical advice for planning
- Modals work for both single-scenario and multi-scenario analysis views
- Version bumped to 3.9.9

+++ src/services/retirement_model.py
@@ -113,8 +113,8 @@ class RetirementModel:
                 (731200, float('inf'), 0.37),
             ]
 
-        total_tax = np.zeros_like(taxable_income)
-        marginal_rate = np.zeros_like(taxable_income)
+        total_tax = np.zeros_like(taxable_income, dtype=float)
+        marginal_rate = np.zeros_like(taxable_income, dtype=float)
 
         for lower, upper, rate in brackets:
             # Income in this bracket
